fix: ui fixes on landing
