+++ packages/web/src/assets/custom.css
@@ -101,6 +101,23 @@ button.fr-link:hover, .tee-router-link:hover {
   background-color: white;
   z-index: 10;
 }
+.tee-home-step-partners-container {
+  align-items: center
+}
+.tee-home-step-partners-img {
+  max-height: 2rem;
+  width: auto;
+  max-width: 8rem;
+}
+.tee-footer-partners-container {
+  background-image: none
+}
+.tee-footer-partners-img {
+  height: auto;
+  max-height: 5.625rem;
+  width: 10rem
+}
+
 .tee-home-img {
   width: 100%;
 }
@@ -113,6 +130,9 @@ button.fr-link:hover, .tee-router-link:hover {
 }
 .tee-text-light {
   color: #929292;
+}
+.tee-text-blue {
+  color: #000091;
 }
 .tee-text-no-result {
   font-size: 2rem;
