+++ packages/web/src/components/home/TeeHomeSteps.vue
@@ -14,9 +14,9 @@
           />
         </div>
         <div class="fr-col fr-col-xs-10 fr-col-sm-12">
-          <h3>
+          <h4>
             {{ c.title }}
-          </h3>
+          </h4>
         </div>
       </div>
     </div>
@@ -30,11 +30,25 @@
       <p>
         {{ c.text }}
       </p>
+      <div
+        v-if="c.logos"
+        class="fr-grid-row fr-sm-hide fr-justify-center fr-mb-2v tee-home-step-partners-container"
+      >
+        <img
+          v-for="logo in c.logos"
+          :key="logo.img"
+          :src="logo.img"
+          :alt="logo.label"
+          class="fr-mx-2v fr-my-2v tee-home-step-partners-img"
+        />
+      </div>
     </div>
   </div>
 </template>
 
 <script setup lang="ts">
+import partners from '@/utils/partners'
+
 const content = [
   {
     title: 'Je renseigne mon profil avec mon SIRET',
@@ -44,12 +58,13 @@ const content = [
   {
     title: 'Je réponds à quelques questions par thématique',
     img: '/images/home/2.svg',
-    text: 'En fonction de vos objectifs et des enjeux de votre entreprise : bâtiment, mobilité, déchets, eau, énergie...'
+    text: 'En fonction de vos objectifs et des enjeux de mon entreprise : bâtiment, mobilité, déchets, eau, énergie...'
   },
   {
     title: 'J’accède aux aides dont mon entreprise peut bénéficier',
     img: '/images/home/3.svg',
-    text: 'Des propositions d’accompagnements et de financements issues de l’ensemble des partenaires : ADEME, Bpifrance, CCI, CMA, etc.'
+    text: 'Des propositions d’accompagnements et de financements issues de l’ensemble des partenaires : ADEME, Bpifrance, CCI, CMA, etc.',
+    logos: partners
   }
 ]
 </script>
