+++ packages/web/src/utils/partners.ts
@@ -1,0 +1,24 @@
+const partners = [
+  {
+    label: 'ADEME',
+    img: '/images/logos/ademe.svg',
+    href: 'https://www.ademe.fr/'
+  },
+  {
+    label: 'Bpifrance',
+    img: '/images/logos/bpi-france.svg',
+    href: 'https://www.bpifrance.fr/'
+  },
+  {
+    label: 'CCI France',
+    img: '/images/logos/cci-france.png',
+    href: 'https://www.cci.fr/'
+  },
+  {
+    label: 'CMA France',
+    img: '/images/logos/cma-france.png',
+    href: 'https://www.artisanat.fr/'
+  }
+]
+
+export default partners
